Refactor messages-util.
